+++ debug/launch/v3-patrick.ts
@@ -5,7 +5,7 @@ import { spfi } from "@pnp/sp";
 import "@pnp/sp/webs";
 import "@pnp/sp/lists";
 import "@pnp/sp/fields";
-import "@pnp/sp/batching";
+import "@pnp/sp/site-scripts";
 import { Web } from "@pnp/sp/webs";
 import { AssignFrom, CopyFrom, getRandomString, TimelinePipe } from "@pnp/core";
 import { RequestRecorderCache } from "../../test/test-recorder.js";
@@ -14,10 +14,6 @@ import { Queryable } from "@pnp/queryable/queryable.js";
 import { indexOf } from "core-js/core/array";
 import { IList } from "@pnp/sp/lists";
 
-import { graphfi } from "@pnp/graph";
-import "@pnp/graph/invitations";
-import "@pnp/graph/users";
-import "@pnp/graph/batching";
 
 declare var process: { exit(code?: number): void };
 
@@ -56,30 +52,22 @@ export async function Example(settings: ITestingSettings) {
 
         // const recordingPath = join("C:/github/@pnp-fork", ".test-recording");
 
-        const graph = graphfi().using(GraphDefault({
+        const sp = spfi("https://318studios.sharepoint.com/sites/dev").using(SPDefault({
             msal: {
-                config: settings.testing.graph.msal.init,
-                scopes: settings.testing.graph.msal.scopes,
+                config: settings.testing.sp.msal.init,
+                scopes: settings.testing.sp.msal.scopes,
             },
         })).using(PnPLogging(LogLevel.Verbose)); 
 
-        // gain a batched instance of the graph
-        const [batchedGraph, execute] = graph.batched();
+        const scripts = await sp.siteScripts.getSiteScripts();
 
-        // we take a reference to the value returned from .users
-        const users = batchedGraph.users;
+        console.log(scripts.length);
 
-        // we invoke it, adding it to the batch (this is a request to /users), it will succeed
-        users();
+        console.log(JSON.stringify(scripts));
 
-        // we invoke it again, because this instance has already been added to the batch, this request will throw an error
-        users();
-
-        // we execute the batch, this promise will resolve
-        await execute();               
-
-        console.log("here!?");
-
+        for (let i = 0; i < scripts.length; i++) {
+            await sp.siteScripts.deleteSiteScript(scripts[i].Id);
+        }
 
         // const list: IList = sp.web.lists.getByTitle("Config");
         // const [batchedListBehavior, execute] = createBatch(list);