+++ CHANGELOG.md
@@ -50,7 +50,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - odataUrlFrom moved to utils folder
   - getParent signature change, path is second param, baseUrl is third param and only supports string
   - removed "core" preset
-  - Improved web and site contructor to handle removing _api/web from any base supplied
+  - Improved web and site contructor to correctly rebase the web/site urls regardless of the url supplied (i.e. create a web from any sp queryable)
 
 - graph:
   - Renamed export "graph" -> "graphfi" with type GraphFI